Fixed the issue with the playback_movie.py script. ffmpeg mode left unchanged

+++ retro/scripts/playback_movie.py
@@ -37,8 +37,9 @@ def playback_movie(emulator, movie, monitor_csv=None, video_file=None, viewer=No
 
     def killprocs(*args, **kwargs):
         ffmpeg_proc.terminate()
-        viewer_proc.terminate()
-        viewer_proc.wait()
+        if viewer:
+            viewer_proc.terminate()
+            viewer_proc.wait()
         raise BrokenPipeError
 
     while movie.step():
